+++ src/main/java/calculator/Calculator.java
@@ -1,0 +1,8 @@
+package calculator;
+
+public class Calculator {
+    public Operand calculator(Operator operator, Operand left, Operand right) {
+        return operator.calculate(left, right);
+    }
+
+}
